cleanup mosaic generation method location and invocation
